fix: minor fix when no port is defined

+++ includes/ServerConfig.hpp
@@ -15,7 +15,7 @@ class ServerConfig {
 		std::map<int, std::string>	errors;
 		std::vector<RouteConfig*>	routes;
 		std::string root;
-		unsigned int	port;
+		unsigned long int	port;
 		std::vector<std::string>	serverNames;
 		std::string	host;
 		size_t	bodyLimit;
